Add ImageUpload API endpoint
This change adds a simple image upload endpoint which supports receiving
a file with desired height and width passed, resizes that and uploads
the result to s3 before retuning the URL for the newly uploaded image.
This provides a fast and convenient way for the frontend to upload an
image and get it processed, even when the image uploaded is large. This
will also then optimize iamges for our use case.

Most image types are supported, including SVG which is uploaded without
modification.

We also now have test cases, only three at the moment, but they test that
an image can be uploaded (doesn't actually put anything on s3 though).

File size is limited to 50 MB. Received files are stored in memory, so
there is a risk of memory exhaustion here, but currently that attack
vector is negligable.