Add .npmrc to fix permissions issue on EB
Due to sharp requiring node-gyp installed as root, we ahve to allow npm
to install stuff as root if it requires it. This shouldn't affect
developer machines much (you will be prompted for password), but on
Elastic Beanstalk this will make sintalling node-gyp actually work. This
is not by any means ideal, but at least fixes the immeditate issue until
Amazon decides to fix this (or node decides to fix it, it's out of our
hands).