Add crop strategy to image upload
An image can be cropped in various ways. Firstly, it can not be cropped,
but instead try to fit inside the bounding box specified by imageHeight
and imageWidth. Else it can eitehr be centre cropped (default), or we
try to find the place with the biggest Shannon entrypy, or the place
with the most skin colours. The caller should be able to decide on the
different strategies.

This can be set by the resizeStrategy parameter in the request body,
which must be one of type:

[north, northeast, east, southeasy, south, sourthwest, west, northwest,
center, centre, entropy, attention, max]